refactor: handler, util 패키지 상위로 이동
